--- frontend/src/metabase/nav/containers/MainNavbar/SidebarItems/SidebarOnboardingSection/analytics.ts
@@ -1,19 +1,0 @@
-import { trackSimpleEvent } from "metabase/lib/analytics";
-
-export const trackAddDataViaCSV = () => {
-  trackSimpleEvent({
-    event: "data_add_via_csv_clicked",
-  });
-};
-
-export const trackAddDataViaDatabase = () => {
-  trackSimpleEvent({
-    event: "data_add_via_db_clicked",
-  });
-};
-
-export const trackOnboardingChecklistOpened = () => {
-  trackSimpleEvent({
-    event: "onboarding_checklist_opened",
-  });
-};
