--- frontend/src/metabase/nav/containers/MainNavbar/SidebarItems/SidebarOnboardingSection/index.ts
@@ -1,1 +1,0 @@
-export * from "./SidebarOnboardingSection";
