+++ frontend/src/metabase/nav/containers/MainNavbar/analytics.ts
@@ -1,0 +1,7 @@
+import { trackSimpleEvent } from "metabase/lib/analytics";
+
+export const trackOnboardingChecklistOpened = () => {
+  trackSimpleEvent({
+    event: "onboarding_checklist_opened",
+  });
+};
